Add styles and fix bug
When the first item is dragged, the currentId state property is not updated yet and a runtime error appears when it can't find a non-existant id.  Add a check on currentId and if there is exactly one question in the list. If valid, we know that there is an attempt at adding the first question.